+++ tests/conftest.py
@@ -28,11 +28,13 @@ async def mock_acreate(prompt, **kwargs):
     result = MagicMock(prompt=prompt)
     for k, v in kwargs.items():
         setattr(result, k, v)
+    return result
 
 
 @pytest.fixture
 def mock_openai_credentials(monkeypatch) -> OpenAICredentials:
-    mock_completion = MagicMock()
-    mock_completion.acreate.side_effect = mock_acreate
-    monkeypatch.setattr("openai.Completion", mock_completion)
-    return OpenAICredentials(api_key="my_api_key", _mock_completion=mock_completion)
+    mock_model = MagicMock()
+    mock_model.acreate.side_effect = mock_acreate
+    monkeypatch.setattr("openai.Completion", mock_model)
+    monkeypatch.setattr("openai.Image", mock_model)
+    return OpenAICredentials(api_key="my_api_key", _mock_model=mock_model)